+++ cmake/Modules/TileDBCommon.cmake
@@ -91,3 +91,25 @@ function(install_target_libs LIB_TARGET)
   endif()
 endfunction()
 
+# Taken from https://github.com/googleapis/google-cloud-cpp/blob/50119a35efc97e92fc20fc6369e72feccdeab984/super/ExternalProjectHelper.cmake#L25
+# Licensed under Apache Licensed v2
+function (set_external_project_build_parallel_level var_name)
+  if ("${CMAKE_GENERATOR}" STREQUAL "Unix Makefiles" OR "${CMAKE_GENERATOR}"
+          STREQUAL "Ninja")
+    if (DEFINED ENV{NCPU})
+      set(${var_name}
+              "--" "-j" "$ENV{NCPU}"
+              PARENT_SCOPE)
+    else ()
+      include(ProcessorCount)
+      processorcount(NCPU)
+      set(${var_name}
+              "--" "-j" "${NCPU}"
+              PARENT_SCOPE)
+    endif ()
+  else ()
+    set(${var_name}
+            ""
+            PARENT_SCOPE)
+  endif ()
+endfunction ()